deps_cross_checker.py: Ensure we cross-check all project combinations.
Old code had a bug: The call to list.remove() effectively removes the
head of the list, but the iterator doesn't realise this.

In effect, we were skipping cross-check for every second item in the list,
but only with projects that come after it in the list.

Replaced iteration with a while loop that *consumes* the project list.